+++ src/main/java/dev/venetsky/operation/TransferAccountCommand.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 import java.util.Scanner;
 
 @Component
-public class TransferAccountCommand implements OperationCommand{
+public class TransferAccountCommand implements OperationCommand {
 
     private final Scanner scanner;
     private final AccountService accountService;
@@ -19,9 +19,9 @@ public class TransferAccountCommand implements OperationCommand{
     @Override
     public void executeCommand() {
         System.out.println("Enter account id to transfer from:");
-        int accountFrom = Integer.parseInt(scanner.nextLine());
+        long accountFrom = Long.parseLong(scanner.nextLine());
         System.out.println("Enter account id to transfer to:");
-        int accountTo = Integer.parseInt(scanner.nextLine());
+        Long accountTo = Long.parseLong(scanner.nextLine());
         System.out.println("Enter money amount to transfer:");
         int money = Integer.parseInt(scanner.nextLine());
         accountService.transferMoney(accountFrom, accountTo, money);
